# backend/client-service/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Parking Client Service API
  description: |
    API for managing clients (subscription owners) and their vehicles.
    This is the central service for registering and identifying subscribers.
  version: 1.0.0
servers:
  - url: http://localhost:8080 # Internal container port
    description: Client Service Development Server
  - url: http://client-service:8080 # Internal Docker network name
    description: Client Service Docker Network

tags:
  - name: Client
    description: Operations with clients and subscribers
    x-interface-name: ClientApi
  - name: Vehicle
    description: Operations with vehicles
    x-interface-name: VehicleApi

paths:
  # =================================================================
  # CLIENTS
  # =================================================================
  /api/clients:
    get:
      tags:
        - Client
      summary: Get a list of all registered clients
      operationId: getAllClients
      responses:
        '200':
          description: Successful response. List of clients.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ClientResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags:
        - Client
      summary: Register a new client
      operationId: createClient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClientRequest'
      responses:
        '201':
          description: Client successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /api/clients/{id}:
    get:
      tags:
        - Client
      summary: Get client by ID
      operationId: getClientById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags:
        - Client
      summary: Update client data by ID
      operationId: updateClient
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClientRequest'
      responses:
        '200':
          description: Client successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '400':
          $ref: '#/components/responses/BadRequestError'

  /api/clients/search:
    get:
      tags:
        - Client
      summary: Search client by phone number
      operationId: searchClientByPhone
      parameters:
        - name: phone
          in: query
          required: true
          schema:
            type: string
            minLength: 1
            example: "+380501234567"
          description: Client phone number (used for quick identification)
      responses:
        '200':
          description: Client found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '404':
          $ref: '#/components/responses/NotFoundError'

  # =================================================================
  # VEHICLES
  # =================================================================
  /api/vehicles:
    get:
      tags:
        - Vehicle
      summary: Get a list of all vehicles
      operationId: getAllVehicles
      responses:
        '200':
          description: Successful response. List of all vehicles.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/VehicleResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags:
        - Vehicle
      summary: Register a new vehicle
      operationId: createVehicle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VehicleCreateRequest'
      responses:
        '201':
          description: Vehicle successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VehicleResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '409':
          description: Conflict - license plate already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /api/vehicles/{id}:
    get:
      tags:
        - Vehicle
      summary: Get vehicle by ID
      operationId: getVehicleById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VehicleResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags:
        - Vehicle
      summary: Update vehicle data by ID
      operationId: updateVehicle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VehicleUpdateRequest'
      responses:
        '200':
          description: Vehicle successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VehicleResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '409':
          description: Conflict - license plate already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    delete:
      tags:
        - Vehicle
      summary: Delete vehicle by ID
      operationId: deleteVehicle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Vehicle successfully deleted
        '404':
          $ref: '#/components/responses/NotFoundError'

  /api/clients/{clientId}/vehicles:
    post:
      tags:
        - Vehicle
      summary: Add a new vehicle to a client
      operationId: addVehicleToClient
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VehicleRequest'
      responses:
        '201':
          description: Vehicle successfully added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VehicleResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '400':
          $ref: '#/components/responses/BadRequestError'
    get:
      tags:
        - Vehicle
      summary: Get a list of all client's vehicles
      operationId: getVehiclesByClient
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: List of vehicles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/VehicleResponse'

components:
  schemas:
    # --- Schemas for Client (DTO) ---
    ClientRequest:
      type: object
      properties:
        fullName:
          type: string
          example: Ihor Petrenko
        phoneNumber:
          type: string
          example: "+380501234567"
        email:
          type: string
          format: email
          example: ihor@example.com
      required:
        - fullName
        - phoneNumber

    ClientResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 101
        fullName:
          type: string
        phoneNumber:
          type: string
        email:
          type: string
        registeredAt:
          type: string
          format: date-time
          example: 2025-11-20T10:00:00Z

    # --- Schemas for Vehicle (DTO) ---
    VehicleCreateRequest:
      type: object
      properties:
        licensePlate:
          type: string
          example: 'AA7777OO'
        clientId:
          type: integer
          format: int64
          example: 101
          description: ID of the client who owns this vehicle
        isAllowed:
          type: boolean
          example: true
          description: Whether the vehicle is allowed to enter
      required:
        - licensePlate
        - clientId

    VehicleUpdateRequest:
      type: object
      properties:
        licensePlate:
          type: string
          example: 'AA7777OO'
        clientId:
          type: integer
          format: int64
          example: 101
        isAllowed:
          type: boolean
          example: true

    VehicleRequest:
      type: object
      properties:
        licensePlate:
          type: string
          example: 'AA7777OO'
      required:
        - licensePlate

    VehicleResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 205
        licensePlate:
          type: string
        isAllowed:
          type: boolean
          example: true
        clientId:
          type: integer
          format: int64

    # --- Schema for Errors ---
    ApiError:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
          example: 2025-11-20T18:30:00Z
        status:
          type: integer
          example: 404
        error:
          type: string
          example: Not Found
        message:
          type: string
          example: Client with ID 123 not found.

  responses:
    NotFoundError:
      description: Resource not found (404)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    BadRequestError:
      description: Bad request or invalid data (400)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    InternalServerError:
      description: Internal server error (500)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'